requirements.txt
boto==2.38.0
Cython==0.23.2
h5py==2.5.0
Keras==0.1.3
mincepie==0.1.0
numpy==1.9.3
Pillow==2.9.0
python-gflags==2.0
PyYAML==3.11
requests==2.8.1
scipy==0.16.0
six==1.9.0
Theano==0.7.0
wheel==0.24.0
